requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pgvector
openai
pydantic
alembic
python-dotenv
requests
python-docx
PyMuPDF
